+++ equinox/experimental/__init__.py
@@ -1,2 +1,2 @@
-from .normalisation import BatchNorm
+from .batch_norm import BatchNorm
 from .stateful import get_state, set_state, StateIndex
