+++ equinox/experimental/batch_norm.py
